+++ Cargo.toml
@@ -143,7 +143,7 @@ version = "49.0.0"
 version = "49.0.0"
 
 [workspace.dependencies.datafusion-sql]
-version = "48.0.0"
+version = "49.0.0"
 default-features = false
 
 [workspace.dependencies.thiserror]
