+++ vegafusion-common/src/error.rs
@@ -65,8 +65,8 @@ pub enum VegaFusionError {
     #[error("Arrow error: {0}\n{1}")]
     ArrowError(ArrowError, ErrorContext),
 
-    #[error("Unparser error: {0}\n{1}")]
-    UnparserError(String, ErrorContext),
+    #[error("Vendor error: {0}\n{1}")]
+    VendorError(String, ErrorContext),
 
     #[error("DataFusion error: {0}\n{1}")]
     DataFusionError(DataFusionError, ErrorContext),
@@ -156,9 +156,9 @@ impl VegaFusionError {
                 context.contexts.push(context_fn().into());
                 VegaFusionError::DataFusionError(err, context)
             }
-            UnparserError(msg, mut context) => {
+            VendorError(msg, mut context) => {
                 context.contexts.push(context_fn().into());
-                VegaFusionError::UnparserError(msg, context)
+                VegaFusionError::VendorError(msg, context)
             }
             #[cfg(feature = "proto")]
             DataFusionProtoError(err, mut context) => {
@@ -207,8 +207,8 @@ impl VegaFusionError {
         }
     }
 
-    pub fn unparser<S: Into<String>>(message: S) -> Self {
-        Self::UnparserError(message.into(), Default::default())
+    pub fn vendor<S: Into<String>>(message: S) -> Self {
+        Self::VendorError(message.into(), Default::default())
     }
 
     pub fn parse<S: Into<String>>(message: S) -> Self {
@@ -270,8 +270,8 @@ impl VegaFusionError {
             DataFusionError(err, context) => {
                 VegaFusionError::ExternalError(err.to_string(), context.clone())
             }
-            UnparserError(msg, context) => {
-                VegaFusionError::UnparserError(msg.clone(), context.clone())
+            VendorError(msg, context) => {
+                VegaFusionError::VendorError(msg.clone(), context.clone())
             }
             #[cfg(feature = "proto")]
             DataFusionProtoError(err, context) => {
