+++ vegafusion-python/Cargo.toml
@@ -10,9 +10,7 @@ name = "vegafusion"
 crate-type = ["cdylib"]
 
 [features]
-default = ["logical-plan"]
 protobuf-src = ["vegafusion-core/protobuf-src"]
-logical-plan = []
 
 [dependencies]
 sysinfo = "0.35.0"
@@ -64,6 +62,9 @@ version = "2.0.4"
 [dependencies.vegafusion-runtime]
 path = "../vegafusion-runtime"
 version = "2.0.4"
+
+[dependencies.datafusion]
+workspace = true
 
 [dependencies.tokio]
 workspace = true
