+++ vegafusion-python/pyproject.toml
@@ -14,7 +14,7 @@ classifiers = [
   "License :: OSI Approved :: BSD License",
   "Topic :: Scientific/Engineering :: Visualization",
 ]
-dependencies = ["arro3-core", "packaging", "narwhals>=1.42"]
+dependencies = ["arro3-core", "packaging", "narwhals>=1.42", "datafusion>=48.0.0,<49.0.0"]
 [[project.authors]]
 name = "VegaFusion Contributors"
 
@@ -32,7 +32,6 @@ Documentation = "https://vegafusion.io"
 
 [project.optional-dependencies]
 embed = []
-datafusion = ["datafusion>=48.0.0,<49.0.0"]
 
 [tool.maturin]
 module-name = "vegafusion._vegafusion"
