+++ vegafusion-runtime/src/task_graph/grpc_runtime.rs
@@ -35,9 +35,6 @@ impl VegaFusionRuntimeTrait for GrpcVegaFusionRuntime {
         indices: &[NodeValueIndex],
         inline_datasets: &HashMap<String, VegaFusionDataset>,
     ) -> Result<Vec<NamedTaskValue>> {
-        // TODO: currently plan executor is not used as there is no way to reliably serialize and transmit it.
-        // Should we just raise an error is fomethign other than None is passed?
-
         let inline_datasets = encode_inline_datasets(inline_datasets)?;
         let request = QueryRequest {
             request: Some(query_request::Request::TaskGraphValues(
